requirements.txt
Django==1.10.6
djangorestframework==3.6.2
drf_nested_routers==0.11.1
django_wysiwyg_redactor==0.5.0
Pillow==4.0.0
gunicorn==19.6.0
newrelic==2.80.1.61
psycopg2==2.7
whitenoise==3.3.0
dj-database-url==0.4.2
simplejson==3.10.0
chardet==2.3.0
python-weixin==0.1.9.1
